# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "H&M Fashion Recommendation API",
    "description": "Personalized fashion recommendations via LightGBM LambdaRank.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/experiment/power": {
      "post": {
        "tags": [
          "experiment"
        ],
        "summary": "Sample size and power calculation for an A/B test",
        "operationId": "power_api_v1_experiment_power_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PowerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PowerResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiment/simulate": {
      "post": {
        "tags": [
          "experiment"
        ],
        "summary": "Run vectorized A/B or A/A simulation using week-104 per-customer arrays",
        "operationId": "simulate_api_v1_experiment_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PowerRequest": {
        "properties": {
          "baseline_rate": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Baseline Rate"
          },
          "relative_lift": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Relative Lift"
          },
          "alpha": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Alpha",
            "default": 0.05
          },
          "power": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Power",
            "default": 0.8
          },
          "weekly_traffic": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Traffic"
          },
          "cuped_variance_reduction": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "minimum": 0.0,
            "title": "Cuped Variance Reduction",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "baseline_rate",
          "relative_lift"
        ],
        "title": "PowerRequest"
      },
      "PowerResult": {
        "properties": {
          "n_per_arm": {
            "type": "integer",
            "title": "N Per Arm"
          },
          "n_per_arm_statsmodels": {
            "type": "integer",
            "title": "N Per Arm Statsmodels"
          },
          "n_per_arm_cuped": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Per Arm Cuped"
          },
          "weeks_needed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weeks Needed"
          },
          "cohen_h": {
            "type": "number",
            "title": "Cohen H"
          }
        },
        "type": "object",
        "required": [
          "n_per_arm",
          "n_per_arm_statsmodels",
          "n_per_arm_cuped",
          "weeks_needed",
          "cohen_h"
        ],
        "title": "PowerResult"
      },
...
      "SimulateRequest": {
        "properties": {
          "test_type": {
            "type": "string",
            "pattern": "^(ab|aa)$",
            "title": "Test Type",
            "default": "ab"
          },
          "n_per_arm": {
            "type": "integer",
            "maximum": 100000.0,
            "exclusiveMinimum": 0.0,
            "title": "N Per Arm"
          },
          "n_sims": {
            "type": "integer",
            "maximum": 2000.0,
            "exclusiveMinimum": 0.0,
            "title": "N Sims",
            "default": 100
          },
          "peeking": {
            "type": "boolean",
            "title": "Peeking",
            "default": false
          },
          "n_days": {
            "type": "integer",
            "maximum": 60.0,
            "exclusiveMinimum": 0.0,
            "title": "N Days",
            "default": 14
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          }
        },
        "type": "object",
        "required": [
          "n_per_arm"
        ],
        "title": "SimulateRequest"
      },
      "SimulateResult": {
        "properties": {
          "test_type": {
            "type": "string",
            "title": "Test Type"
          },
          "n_per_arm": {
            "type": "integer",
            "title": "N Per Arm"
          },
          "n_sims": {
            "type": "integer",
            "title": "N Sims"
          },
          "false_positive_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "False Positive Rate"
          },
          "empirical_power": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Empirical Power"
          },
          "peeking_fpr": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peeking Fpr"
          }
        },
        "type": "object",
        "required": [
          "test_type",
          "n_per_arm",
          "n_sims",
          "false_positive_rate",
          "empirical_power",
          "peeking_fpr"
        ],
        "title": "SimulateResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}